storeMS/requirements.txt
Django>=3.0,<4.0
mysqlclient
python-dotenv==1.0.1
djangorestframework
requests
python-dotenv
django-autoslug
python-jose
django-cors-headers
drf-yasg
setuptools